Moved dwarf type parsing into anal/type_dwarf.c
