+++ libr/anal/meson.build
@@ -33,6 +33,7 @@ r_anal_sources = [
   'sign.c',
   'switch.c',
   'type.c',
+  'type_dwarf.c',
   'value.c',
   'var.c',
   'vtable.c',
