+++ libr/bin/meson.build
@@ -142,7 +142,6 @@ r_bin = library('r_bin', r_bin_sources,
   include_directories: r_bin_inc,
   c_args: ['-DR_API_BIN_ONLY=1'] + library_cflags,
   dependencies: [
-    r_anal_dep,
     r_util_dep,
     r_io_dep,
     r_hash_dep,
